Augumented brain images
Every data collection entry contains the required patient data. The goal of the dataset is to analytically forecast if a patient would have a BS based on particular recalled indicative estimations. Consequently, the dataset used in the experiment does not contain any risk of patient confidentiality. In the dataset, there are 5110 samples. Each sample contains eleven attributes and one target value.